mobile/Paytm_hackathon-main/backend/requirements.txt
fastapi[standard]
uvicorn
python-multipart
python-dotenv
motor
dnspython
PyJWT
passlib[bcrypt]
bcrypt
httpx
groq
numpy
librosa
faster-whisper
resemblyzer
scikit-learn
speechbrain
spacy
supabase
xgboost
pyannote.audio
opencv-python-headless
